pluginify change handling
